gfs2-utils: Fixed patches for updated GFS2
Removed old, for the actual gfs2 version (3.1.8), unnecessary patch.
Made patch for gfs2 lockcapture, so that it uses corosync tools directly.
We do not throw an error when no GFS2 entry is in /etc/fstab, because
there cannot possible be an entry after the first installation and it's
very confusing to get an error at this point. A info message should be
enough. Using gfs2-utils.git tag 3.1.8 for this build.

Signed-off-by: Thomas Lamprecht <t.lamprecht@proxmox.com>